requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
billiard==3.6.4.0
cached-property==1.5.2
cachetools==4.2.2
celery==5.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.3
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.0.1
joblib==1.0.1
kombu==5.1.0
nltk==3.6.2
oauthlib==3.1.0
packaging==20.9
pipdeptree==2.0.0
prompt-toolkit==3.0.18
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
tqdm==4.61.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
